Pass arguments through environment variables
The coordinator modifies the communication config by adding an randomly generated prefix to all topics on startup. This prefix is unique to the instance and not written back to the dataflow file. Since the runtime nodes need to use that exact same prefix, they need to be started using the modified communication config, instead of reading the original config from the dataflow file again.

This commit changes the argument passing to environment variables, which makes it possible to use the modified communication config. It also makes the runtime nodes more similar to custom nodes, which also get their arguments through env variables.