Create general `communication::init` function to abstract CommunicationLayer creation
The function derives the communication layer type from the config and handles all the initialization. This commit also updates the format of the communication config to better differentiate the different layer types.

To make the communication code more general, we also abstract away the topic prefixing into the communication layer. The publish and subscribe functions will now add the prefix automatically.