Create a C-API for operators and an example operator
Adds an `api.h` header file that defines the dora functions in a C-compatible way. To showcase that C-operators can interact with Rust operators and custom nodes, we create a simple C operator and add it to our `mini-dataflow`. The operator reads the periodic timestamps, increments a counter, and sends the counter value out. The counter is kept across calls by storing it behind the context pointer.