Generate a standard message format for node and operator
This commit forces the use of a standard structure for both node and operator.

The message needs to have a metadata and a data section.

The metadata will contain every information not related to the main data payload.

The information about the topic from which the message is received or sent to.

The information about the open telemetry context and in the future about the watermark.