Make `PublishSample` an owned type and change output API to 3-step process
Instead of passing a closure, we now `prepare` the output first, then write the data, and finally send it out. This should make it possible to enable zero-copy send operations also in Rust/C operators and C nodes since we avoid the type and lifetime complexity that an additional closure argument would cause.