Changing `drop_operator` to `__del__` in `python.rs`
Using the default python destructor function for dora operator simplify
the usability of dora as user does not have to learn as many custom
methods.

This also creates a symmetry with __init__ and  reduce the risk of
errors if drop_operator is not present or misspelled.

It also enables the usability of dora operator outside of dora ( for
testing dor example ) and let it be naturally garbage collected.