Only drop input events and free shmem on dropping
We don't want to drop other types of events (e.g. stop or input closed) because they might be important for the correct functionality. For this reason, we only drop input events now. To implement this, we use a FIFO buffer in the listener instead of dropping the events on the sender side.

This commit also fixes a memory leak that occured when dropping values. We now report the drop tokens properly to the shared memory handler when dropping an input.