Await all nodes to finish before marking dataflow as finished
Previouly, if one node failed we would immidiately mark the dataflow as
finished even though other nodes might still be running.

This meant that if one node failed, it would not be possible to stop the
dataflow through the cli as it is marked as finished.

This commit will just warn the user that one node failed but will keep
await that all nodes within the dataflow finishes.