Fix looping in daemon listener loop
The event channel might be closed already inside the loop. When using a normal `recv()` then, we got a future that immediately resolved to `None` and let to lots of unnecessary loop iterations, or even endless loops. This commit fixes this by using a manually constructed future instead, which will remain pending forever once the event channel is closed. This way, we only wait for the next node message in the listener loop.