Simplify parsing of `AMENT_PREFIX_PATH`
This PR simplify how ROS2 messages are parsed using dora.

It will parse all messages included in the `AMENT_PREFIX_PATH` folder
that has a path in the following format: `<package>/msg/<message>.msg`.

Similarly for services it will be parsed as: `<package>/srv/<service>.srv`.

It does not requires `.idl` anymore nor does it needs to be in the `ament_index`.