Fix: Wait until dora daemon is connected to coordinator on `dora up`
It might take a short time until the daemon is connected to the coordinator after start-up. This could lead to issues when using `dora up` in a script because a subsequent `dora start` command might fail when the daemon is not connected at that point.

This commit fixes this issue by waiting until the coordinator confirms that a daemon is connected before returning from `dora up`.