Forbid `/` in node IDs
We use `/` as separator for input mappings in dataflow YAML files (e.g. input: `node/output_id`).

We do support `/` characters in output IDs for runtime nodes. If we allowed `/` characters in node IDs too, splitting a `a/b/c` mapping would no longer be possible (could be node `a` with output `b/c` or node `a/b` with output `c`). Therefore we explicitly error on node IDs containing `/` characters in this commit.