Simplify UNC working directory paths on Windows
Fixes the tests/examples of this repo. A UNC path as working directory doesn't work because the build commands of our examples invoke `cargo` on the dora repo. The dora repo uses wildcards in the member list in the top-level Cargo.toml, which are ignored in UNC-paths. Thus, `failed to load manifest for workspace member` errors occur.