Spawn, clone, and build nodes sequentially
Spawn nodes one-by-one instead of doing it concurrently. Subsequent nodes might want to reuse git repos or build artifacts of previous nodes. Spawning them in parallel might lead to conflicts and errors (e.g. because a git repo reuse is handled before a git clone or because some build command is not concurrency-safe).