Error if node with `git` field was not built before dataflow start
The working directory of nodes with git fields is only decided on `dora build` because that's where we translate the branch/tag name into the commit hash.

Before, we fell back to using the base working directory if no build was done. However, this is confusing since the reason for the different working directory is not communicated in any.

This commit throws an error if no build happened before starting a dataflow with git nodes.