Make nodes runnable standalone, prompting the user for inputs
Adds a new interactive mode to the Rust node API, which automatically activates when no `DORA_NODE_CONFIG` env variable is set. When run in interactive mode, the user is prompted for inputs on the command line. Input data can be given as string or as JSON.

This feature makes it possible to try a node with some inputs and see how it reacts, without the need to set up a full dataflow for it. This can be useful during development.