Create cargo update PR as draft
PRs created through the GITHUB_TOKEN don't trigger subsequent actions, so the CI doesn't run on the created PR. As a workaround, we create the PR as draft and use the new `ready_for_review` trigger to run the CI (when we mark it as ready manually). I cc'ed myself to be still notified about the created draft PRs.