Only run in interactive mode if `stdin` is a terminal
We have some Python tests that try to initialize a dora node. We don't want to run in interactive mode in this case because the executable is run by `pytest`. So there is no way to supply interactive input. By checking if stdin is a terminal, we can continue to throw an error in this case.