Use `now_or_never` instead of `try_recv`
Constructing the `Stream` on the fly in `poll_next` doesn't work as dropping the stream will also drop the registered waker handle. So no wakeup happens anymore.

So we need to keep the long-lived `RecvStream` type, which doesn't have a `try_recv` method. However, we can use `next().now_or_never()` for the same effect.