Don't use the scheduler for integration testing
It leads to a non-deterministic event order because non-input events are always delivered first. So the event order depends on the number of queued events.

It also specifies a fixed queue size of 1 for inputs that are not known before. For integration testing we don't want to use a limited queue size.