Remove custom `Deserialize` implementation for `LogLevelOrStdout`
To avoid similar errors in the future. By deriving the traits, we can be sure that the `Serialize` and `Deserialize` impls are always compatible.

Note that the `Deserialize` impl for `LogLevel` [1] is calling the `FromStr` implementation [2], which ignores case. So deserializing 'INFO', 'info', and 'InFO' will all still work.

The `stdout` variant needs to match exactly, but this is a custom extension of ours, so we can ensure to use lowercase everywhere.

[1]: https://docs.rs/log/latest/src/log/serde.rs.html#61
[2]: https://docs.rs/log/latest/src/log/lib.rs.html#517-525